Move Put and TestDBPut to their own files
